Restructured debugger
