+++ services/syra-daw-fe/src/App.tsx
@@ -4,16 +4,8 @@ import { useHotkeys } from 'react-hotkeys-hook';
 import Debugger from './ui/debug/Debugger';
 import { FpsView } from 'react-fps';
 import {
-  Button,
   ChakraProvider,
   extendTheme,
-  Modal,
-  ModalBody,
-  ModalCloseButton,
-  ModalContent,
-  ModalFooter,
-  ModalHeader,
-  ModalOverlay,
 } from '@chakra-ui/react';
 import useListenForEngineStart from "./hooks/tone/useListenForEngineStart";
 
@@ -28,8 +20,6 @@ function App() {
   const [showDebugMenu, setShowDebugMenu] = useState(false);
   const [showFpsMeter, setShowFpsMeter] = useState(false);
 
-  const onCloseDebugger = () => setShowDebugMenu(false);
-
   useHotkeys('shift+d', () => setShowDebugMenu((currVal) => !currVal));
   useHotkeys('shift+f', () => setShowFpsMeter((currVal) => !currVal));
 
@@ -40,23 +30,7 @@ function App() {
       <ChakraProvider theme={customTheme}>
         <AppRouter />
         {showFpsMeter && <FpsView />}
-
-        <Modal isOpen={showDebugMenu} onClose={() => setShowDebugMenu(false)} size={'6xl'}>
-          <ModalOverlay />
-          <ModalContent>
-            <ModalHeader>Debugger</ModalHeader>
-            <ModalCloseButton />
-            <ModalBody>
-              <Debugger />
-            </ModalBody>
-
-            <ModalFooter>
-              <Button colorScheme="blue" mr={3} onClick={onCloseDebugger}>
-                Close
-              </Button>
-            </ModalFooter>
-          </ModalContent>
-        </Modal>
+        {process.env.NODE_ENV === 'development' && <Debugger onClose={() => setShowDebugMenu(false)} showMenu={showDebugMenu}/>}
       </ChakraProvider>
     </StrictMode>
   );
