+++ services/syra-daw-fe/syra-daw-fe.yml
@@ -14,7 +14,7 @@ spec:
     spec:
       containers:
         - name: syra-daw-fe
-          image: registry.digitalocean.com/syra-do/services/syra-daw-fe:v0.0.13
+          image: registry.digitalocean.com/syra-do/services/syra-daw-fe:v0.0.14
           imagePullPolicy: Always
           ports:
             - containerPort: 80
